Move test files to test/
