+++ test/w.js
@@ -1,7 +1,7 @@
 import test from 'ava'
 import express from 'express'
 import request from 'supertest'
-import w from '../w.js'
+import w from '../lib/w.js'
 
 test('w with a simple 200 response', async t => {
   const app = express()
